+++ js/es6/vendors/index.js
@@ -6,6 +6,7 @@ import './jquery.hammer';
 import './../../../bower_components/jquery-elementresize/dist/jquery.elementresize';
 import './../../../bower_components/devbridge-autocomplete/dist/jquery.autocomplete';
 window.Dropzone = global.Dropzone = require('./../../../bower_components/dropzone/dist/dropzone');
+window.toastr = global.toastr = require('./../../../bower_components/toastr/toastr');
 window.autosize = global.autosize = require('./../../../bower_components/autosize/dist/autosize');
 import './../../../bower_components/velocity/velocity';
 import './../../../bower_components/velocity/velocity.ui';